% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-viz.R
\name{as_df}
\alias{as_df}
\title{Turn the node list into a data frame of attributes}
\usage{
as_df(x)
}
\value{
a data frame of id's and node attributes
}
\description{
Turn the node list into a data frame of attributes
}
