% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-viz.R
\name{as_edges}
\alias{as_edges}
\title{Take a node list and return a list of from/to edges identified by id}
\usage{
as_edges(x)
}
\arguments{
\item{x}{node list}
}
\value{
a list (one per node) of lists (one per edge) of
  list(from = <FROM_ID> to = <TO_ID>)
}
\description{
A helper function, returns a structured list
}
