% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-viz.R
\name{as_graph}
\alias{as_graph}
\title{Turn the node list into a graph object}
\usage{
as_graph(x)
}
\arguments{
\item{x}{node list}
}
\value{
igraph graph representing the node-list and parent-child
  relationships.
}
\description{
Currently uses igraph
}
