% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R, R/utils.R
\name{BICs}
\alias{BICs}
\title{calculate the BIC value}
\usage{
BICs(xi, beta, th, mu, update)
}
\value{
the BIC value
}
\description{
calculate bic
}
\details{
calculate the BIC value of the current model
}
