% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-node.R
\name{compose_id}
\alias{compose_id}
\title{Create a numeric id from a name and environment}
\usage{
compose_id(name, env, generator = smgr::default_hash())
}
\arguments{
\item{name}{name of object the id is created for}

\item{env}{an env with object contents}
}
\value{
a string id (hash)
}
\description{
All components are combined, none are ignored, to create the id.
}
