% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-node.R
\name{compose_tag}
\alias{compose_tag}
\title{Create a label ("tag") from a name and environment}
\usage{
compose_tag(name, env, sep = smgr:::default_sep())
}
\arguments{
\item{name}{name of object the tag is created for}

\item{env}{an env with object contents}

\item{sep}{separator to use in the tag}
}
\value{
a string tag
}
\description{
All components are combined, none are ignored, to create the tag.
}
