% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_unique}
\alias{count_unique}
\title{Count unique values in x}
\usage{
count_unique(x)
}
\arguments{
\item{x}{object to check}
}
\value{
count of unique values in x
}
\description{
Count unique values in x
}
