% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_hash}
\alias{default_hash}
\title{md5 is fine for a collision-resistant identifier}
\usage{
default_hash(x)
}
\arguments{
\item{x}{thing to hash}
}
\value{
a hash string
}
\description{
Some day we should add checking and bump the hash
complexity on collision (?)
}
