% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge.R
\name{DirectedEdge}
\alias{DirectedEdge}
\title{R6 class representing the directed edge between two nodes}
\description{
R6 class representing the directed edge between two nodes

R6 class representing the directed edge between two nodes
}
\details{
Edges in graph building are implicit (stored as child nodes for 
each node) so these edges are used primarily for implementing 
simulation.  They store references to the parent nodes and 
they can modfiy both nodes.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DirectedEdge$new()}}
\item \href{#method-do}{\code{DirectedEdge$do()}}
\item \href{#method-transfer}{\code{DirectedEdge$transfer()}}
\item \href{#method-clone}{\code{DirectedEdge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Used as DirectedEdge$new(...) create a new edge
  between two nodes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectedEdge$new(from, tail, to, head)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{(directed) edge originates at this node}

\item{\code{tail}}{effect of edge on the ('from') node it originates
from as a list of `dplyr::mutate`-style expressions.  The 
expressions can access the special objects `.from` and `.to` 
which are used to access the nodes at the head or tail of 
the edge.  They are used as `.from$foo` to access attribute or
or data `foo` in the tail node.}

\item{\code{to}}{(directed) edge goes to this node}

\item{\code{head}}{parallel to `tail`, but for the `to` node.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a 'DirectedEdge' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do"></a>}}
\if{latex}{\out{\hypertarget{method-do}{}}}
\subsection{Method \code{do()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectedEdge$do(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transfer"></a>}}
\if{latex}{\out{\hypertarget{method-transfer}{}}}
\subsection{Method \code{transfer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectedEdge$transfer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DirectedEdge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
