% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Evaluate}
\alias{Evaluate}
\title{calculate the Dunn index and Sihouette index}
\usage{
Evaluate(mat, cluster)
}
\arguments{
\item{mat}{input data}

\item{cluster}{identified programs}
}
\description{
calculate the Dunn index and Sihouette index
}
