% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_name}
\alias{has_name}
\title{Check if object has a specific name}
\usage{
has_name(x, name)
}
\arguments{
\item{x}{object to check on}

\item{name}{name to check for}
}
\value{
TRUE iff `name` is a name in `x`
}
\description{
Check if object has a specific name
}
