% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R, R/utils.R
\name{loglik}
\alias{loglik}
\title{calculate the loglikehood}
\usage{
loglik(th, mu, y)
}
\arguments{
\item{th}{theta in the negative binomial distribution}

\item{mu}{mu in the negative binomial distribution}
}
\value{
A numeric vector of log density.
}
\description{
loglikehood
}
\details{
calculate the loglikehood based on the negative binomial or zinb distribution
}
