% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{max_nchar}
\alias{max_nchar}
\title{Length of longest character string in x}
\usage{
max_nchar(x)
}
\arguments{
\item{x}{object to check}
}
\value{
length of longest character string in x
}
\description{
Length of longest character string in x
}
