% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{Process}
\alias{Process}
\title{R6 class representing a process (a list of transitions}
\description{
R6 class representing a process (a list of transitions

R6 class representing a process (a list of transitions
}
\details{
Not much here but a fancy list...
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{size}}{the number of transitions in the process}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Process$new()}}
\item \href{#method-insert}{\code{Process$insert()}}
\item \href{#method-get_transition}{\code{Process$get_transition()}}
\item \href{#method-clone}{\code{Process$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Used as Process$new(...) to create a new process from a set of
transtions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{transitions combined to create the process}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a 'Process' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert"></a>}}
\if{latex}{\out{\hypertarget{method-insert}{}}}
\subsection{Method \code{insert()}}{
insert a new transition into the process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$insert(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{transition to insert}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_transition"></a>}}
\if{latex}{\out{\hypertarget{method-get_transition}{}}}
\subsection{Method \code{get_transition()}}{
retrieve a transition by index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$get_transition(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{index of the transition to retrieve}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a transition
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
