% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{short_form}
\alias{short_form}
\title{Find a short version of an identifier that maintains
uniqueness within the vector}
\usage{
short_form(x)
}
\arguments{
\item{x}{object}
}
\value{
truncate all elements to shortest possible string 
        without loosing uniqueness
}
\description{
Find a short version of an identifier that maintains
uniqueness within the vector
}
