% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smgr_main}
\alias{smgr_main}
\title{SMGR main function}
\usage{
smgr_main(
  sm.data,
  K,
  R = 3,
  n_burnin = 200,
  n_draw = 200,
  n_iter = 20,
  option = "nb",
  type = "simulate",
  output = "latent"
)
}
\arguments{
\item{sm.data}{consists of single-cell RNA-seq data and single-cell ATAC-seq data in the format of matrix. Columns represent cells and rows represent genes (scATAC-seq input has been converted to gene-based data matrix)}

\item{K}{The number of features in the single-cell multi-omics data}

\item{R}{the dimension of latent representation}

\item{n_burnin}{burnin for MCMC sampling}

\item{n_draw}{drawing for MCMC sampling}

\item{n_iter}{number of iterations}

\item{option}{nb or zinb}

\item{type}{simulate or experimental data}
}
\value{
latent representation; co-regulation programs
}
\description{
identify co-regulation programs across single-cell multi-omics data
}
