% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.R
\name{Transition}
\alias{Transition}
\title{R6 class representing a single transition}
\description{
R6 class representing a single transition

R6 class representing a single transition
}
\details{
The transition has a process name, conditions it occurs under, and
transformations that it applies to a node

The condition ("match") are specified as in `dplyr::filter` using the 
node attributes to match against.  The transformations follow the 
same approach as `dplyr::mutate` where a data mask basd on the attribute
names is searched prior to checking for varaible names in the quosure
environments.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{of the transition}

\item{\code{process}}{process name stored by the transition}

\item{\code{match}}{process match stored by the transition}

\item{\code{transformation}}{process transformation stored by the transition}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Transition$new()}}
\item \href{#method-do}{\code{Transition$do()}}
\item \href{#method-clone}{\code{Transition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Used as Transition$new(...) create a new transition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transition$new(process, match, transformation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{process}}{name of the process described by the transformation}

\item{\code{match}}{list of conditions that the process applies to}

\item{\code{transformation}}{list of transformations to use on the process}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a 'Transition' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do"></a>}}
\if{latex}{\out{\hypertarget{method-do}{}}}
\subsection{Method \code{do()}}{
Apply this transition to a node if the node matches
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transition$do(x, mutate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a 'Node' object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a child 'Node' object created by applying the transition, if applicable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Transition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
