% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R, R/utils.R
\name{update_Ys}
\alias{update_Ys}
\title{MCMC sampling for identifying latent representation}
\usage{
update_Ys(mean_x, update_y, estimate_y, k, r, n, sdev = 0.05, nburnin, ndraw)
}
\arguments{
\item{mean_x}{A non-negative integer or vector}
}
\value{
updated latent representation
}
\description{
MCMC sampling for identifying latent representation
}
